# Revert [SPARK-10399] [SPARK-23879] [SPARK-23762] [SPARK-25317]