… common/unsafe/src/main/java/org/apache/spark/unsafe/memory/UnsafeMemoryAllocator.java
… common/unsafe/src/main/java/org/apache/spark/unsafe/types/UTF8String.java
… common/unsafe/src/test/java/org/apache/spark/unsafe/PlatformUtilSuite.java
… common/unsafe/src/test/java/org/apache/spark/unsafe/array/LongArraySuite.java
… common/unsafe/src/test/java/org/apache/spark/unsafe/hash/Murmur3_x86_32Suite.java
… common/unsafe/src/test/java/org/apache/spark/unsafe/memory/MemoryBlockSuite.java
… common/unsafe/src/test/java/org/apache/spark/unsafe/types/UTF8StringSuite.java
… core/src/main/java/org/apache/spark/memory/TaskMemoryManager.java
… core/src/main/java/org/apache/spark/shuffle/sort/ShuffleInMemorySorter.java
… core/src/main/java/org/apache/spark/shuffle/sort/ShuffleSortDataFormat.java
… core/src/main/java/org/apache/spark/util/collection/unsafe/sort/UnsafeInMemorySorter.java
… core/src/test/java/org/apache/spark/memory/TaskMemoryManagerSuite.java
… core/src/test/scala/org/apache/spark/util/collection/ExternalSorterSuite.scala
… core/src/test/scala/org/apache/spark/util/collection/unsafe/sort/RadixSortSuite.scala
… mllib/src/main/scala/org/apache/spark/ml/feature/FeatureHasher.scala
… mllib/src/main/scala/org/apache/spark/mllib/feature/HashingTF.scala
… sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/UnsafeArrayData.java
… sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/UnsafeRow.java
… sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/codegen/UTF8StringBuilder.java
… sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/hash.scala
… sql/catalyst/src/test/java/org/apache/spark/sql/catalyst/expressions/HiveHasherSuite.java
… sql/catalyst/src/test/java/org/apache/spark/sql/catalyst/expressions/XXH64Suite.java
… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/codegen/UTF8StringBuilderSuite.scala
… sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OffHeapColumnVector.java
… sql/core/src/main/java/org/apache/spark/sql/vectorized/ArrowColumnVector.java
… sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/SortBenchmark.scala
… sql/core/src/test/scala/org/apache/spark/sql/execution/python/RowQueueSuite.scala