## What changes were proposed in this pull request?

When running TPC-DS benchmarks on 2.4 release, @npoggi and @winglungngai  saw more than 10% performance regression on the following queries: q67, q24a and q24b. After we applying the PR https://github.com/apache/spark/pull/22338, the performance regression still exists. If we revert the changes in https://github.com/apache/spark/pull/19222, @npoggi and @winglungngai  found the performance regression was resolved. Thus, this PR is to revert the related changes for unblocking the 2.4 release. 

In the future release, we still can continue the investigation and find out the root cause of the regression. 

## How was this patch tested?

The existing test cases