@kiszk 2018-09-08
@gatorsmile Overall, I agree with revert since performance degradation is confirmed.

When I run the TPC-DS in #19222, I have not seen such a performance regression as [here](https://github.com/apache/spark/pull/19222#issuecomment-378154854). For future analysis, I am very interested in 
* execution time of each query
* environment (machine, os, jvm, cluster or local, ...)
* scaling factor,
* # of repeats per query
* others

@npoggi and @winglungngai Thanks for your investigations. Would it be possible to share the avobe information?