@HyukjinKwon 2018-09-25
It would be great if the information requested https://github.com/apache/spark/pull/22361#issuecomment-419607024 is shared so that other people can check it as well.