@npoggi 2018-09-25
Hi @kiszk and @HyukjinKwon, first sorry for the late reply, I was trying to get some detailed profiling info for this.  Basically, the regressions in TPCDS SF 1,000 are as follows when compared branch- `2.3` to `2.4` using a 10-workers cluster of 4-vcpus each :
```
q67: from 323.2s to 283.6s (18.7%)
q24a: from 139.1s to 153.9 (12.4%)
q24b: from 128.7s to 145.7s (11.9%)
```
These 3 queries in order are the easiest to root-cause to the new MemoryBlock API, as there is more time spent in UTF8String (getBytes, init, UnsafeRow.getUTF8String).  There were also other possibly affected queries: q8, q11, q54. 
We can test future performance improvements to the proposal, just ping @winglungngai, @cloud-fan, and me in the PR.  Thanks,