@gatorsmile 2018-09-27
Nico will give a talk in the upcoming spark summit: https://databricks.com/session/a-framework-for-evaluating-the-performance-and-the-correctness-of-the-spark-sql-engine 

If possible, please join the talk! : )