@kiszk 2018-09-27
It looks very intersting talk. Since I cannot join the SAIS unfortunally, I will watch this in live stream.